This is what I have set up for testing the server stuff I am working on.
The general idea of the callback routines should be fine.  I will document it more when i get a chance... and fix it.

Right now, it works up to the control one, but then when it starts the send file thread, the socket fails on the connect.  Also, my reading in of the file isn't working properly

Signed-off-by: jbayntun <j_bayntun@hotmail.com>